# Use an sqlite database for the media cache