--- a/src/defaultsettings.cpp
+++ b/src/defaultsettings.cpp
@@ -61,6 +61,8 @@ void set_default_settings()
 	settings->setDefault("curl_file_download_timeout", "300000");
 	settings->setDefault("curl_verify_cert", "true");
 	settings->setDefault("enable_remote_media_server", "true");
+	settings->setDefault("media_cache_class_expiry_time_unknown", "0 s");
+	settings->setDefault("media_cache_class_expiry_time_legacy", "never");
 	settings->setDefault("enable_client_modding", "false");
 	settings->setDefault("max_out_chat_queue_size", "20");
 	settings->setDefault("pause_on_lost_focus", "false");