… builtin/settingtypes.txt
… src/client/CMakeLists.txt
… src/client/filecache.cpp
… src/database/database-sqlite3.h
… src/threading/thread.cpp
… src/unittest/test_utilities.cpp
… src/util/CMakeLists.txt
… src/util/Optional.h
… src/util/basic_macros.h
… src/util/thread.h
… src/util/time_parsing.cpp
… src/util/time_parsing.h