@nerzhul 2021-08-23
whereas i understand the features you want to achieve, storing blobs in sqlite is totally counter productive. I suggest instead that to have cache metadata inside but not the files