@Desour 2021-08-23
>storing blobs in sqlite is totally counter productive

Could you elaborate more on this please? I don't see what is wrong with storing blobs in sqlite.