@SmallJoker 2021-08-23
Isn't this a bit over-engineered? Keep the date of the last - say - 10 server joins and update the modification date of media files when they're used.

Afterwards delete files when `(oldest_server_join_date - cache_file_modification_date) > 30 days`. This way, casual players will not lose the cache right away, and the cache is earliest deleted after 30 days (subject to change) when not visiting a certain server.