@Desour 2021-08-23
SmallJoker:
The cache-classes are there because #10598 suggested it:
>If there are time-based heuristics, the time scale used should probably depend on media type: code and translations will probably have far shorter lifespans than textures or models.

Removing the settings and using fixed times would of course also simplify it a little bit. But imo the settings don't add that much complexity.
Apart from those two points, I think your suggestion is more complicated, because I'd had to store the server join dates.
I don't dislike the suggestion though.