@celeron55 2021-08-23
Has someone benchmarked performance compared to the old filesystem approach?