@Desour 2021-08-23
I haven't benchmarked the whole connection process, but at least the initial step in the media downloader. And I did only test in singleplayer with mtg. I also haven't tested with a very full media cache.

At first the media loading was much slower than with the old filesystem approach, because I've done one sql transaction per file (it took about 7 s). So, I've changed that (see 3rd commit), and now the cache loading only takes about 50 ms on my machine in mtg. This is nearly nothing compared to the ca. 0.4 seconds that the client takes for loading the files into irrlicht stuff or whatever. The benchmark code for that is still there.
The storing of new media files ("update()") was done for each file as it's received. This was also noticeably slow with the one transaction per file approach, so I've also fixed that by storing the file data in ram and updating the cache with one transaction when the media downloader is finished. I haven't measured the time here. However, it doesn't seem like the media cache is a bottleneck here.