@celeron55 2021-08-24
Well, anyway - the problem with the simple timestamp + expiry time approach for deletion is, like discussed in #10598, that if a user stops playing for however long the cache expiry time is, the client shouldn't clear the cache before the user has actually used the client again on multiple servers so that the client actually knows which files are not needed.

One solution to such irregular usage patterns is a two-stage expiration logic, where reaching the expiration time marks the file as "expiration pending", and resets the timestamp, and once the expiration time is reached again with the pending flag already set, then the file will be dropped from the cache. Basically, a file has to expire twice before it's dropped.

I don't agree with nerzhul about storing blobs in sqlite being counter productive. It goes against traditional database design guidelines, but sqlite and many other modern databases are perfectly capable of doing this.