@Desour 2021-08-24
FYI, I've done some more benchmarking:
- I've used a more filled media cache (ca. 350 MiB, 43041 files, data size per file from 34 B to ca. 7.3 MB, avg. filesize 8.2 KB, median file size 514 B). See title comment (How To Test section) for a script to insert your files into the sqlite db.
- I've still just tested in mtg (plus mesecons modpack).
- I've also measured the cache update with new files and the cache cleaning.
- I'm using an ext4 file system.
- The results are not meant to be precise, but just give an idea about the order of magnitude.
- Results:
  - (The client media loading (the `Client::checkAndLoad` stuff) takes about 0.75 seconds, idk why I got 0.4 seconds earlier, maybe I misread.) 
  - The loading from cache is still just about 50 ms, with the filesystem approach and the sqlite db, no matter if I take the sparsely or fatly filled cache. (This is very inaccurate for the filesystem approach, because the measurement there is mixed up with the client media loading, which takes much longer.)
  - The cache update also takes *about* 50 ms if the new files are not yet cached (if all were found, it doesn't do anything). Updating cache and the clean check (without deleting anything) take longer with a fatly filled (update: 34 ms vs. 68 ms, clean check: 1 ms vs. 54 ms). I haven't tested the filesystem approach with a filled up folder (because deleting the ones to be loaded from it would make too much work for me now); with an empty folder it takes 50 ms.
  - TL;DR for the above 2 points: I didn't find a relevant slow-down.
  - However, if I use a filled up db and set the expire time to 0 s (= delete everything), it takes about 10 s to delete. If I just delete everything from the table in the sqlite3 cli (with `delete from files where true;`), it also takes several seconds. So, it might make sense to do the cache cleaning in a separate thread.

>a two-stage expiration logic

Ah, yes. That's a good and simple idea. I'll implement that.