@erlehmann 2021-09-13
I think this is an extremely overengineered solution for the goal of media cache expiration.

By which I mean, it is probably among the worst solutions for a problem that I have seen so far.

Cache expiration and cleanup can easily be done by reading `atime` from the existing cache folder.