@erlehmann 2021-09-13
Note that this also breaks all those cool shell tricks that were used so far, like finding out the filetypes in the cache using `ls |xargs -n50 file --mime --brief |sort |uniq -c`.